+++ database/migrations/2025_07_26_225235_add_entry_channel_to_inscriptions_table.php
@@ -12,8 +12,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('inscriptions', function (Blueprint $table) {
-            $table->unsignedBigInteger('entry_channel')->nullable()->change();
-            $table->foreign('entry_channel')->references('id')->on('entry_channels')->onDelete('set null');
+            if (!Schema::hasColumn('inscriptions', 'entry_channel')) {
+                $table->unsignedBigInteger('entry_channel')->nullable();
+            }
+        });
+        Schema::table('inscriptions', function (Blueprint $table) {
+            if (Schema::hasColumn('inscriptions', 'entry_channel')) {
+                $table->unsignedBigInteger('entry_channel')->nullable()->change();
+                $table->foreign('entry_channel')->references('id')->on('entry_channels')->onDelete('set null');
+            }
         });
     }
